+++ Zero/src/Zero/Renderer/Framebuffer.h
@@ -14,7 +14,7 @@ public:
   virtual ~Framebuffer() = default;
   virtual void Bind() const = 0;
   virtual void Unbind() const = 0;
-  virtual void Resize(const glm::vec2& size) = 0;
+  virtual void Resize(const glm::uvec2& size) = 0;
   virtual uint32_t GetColorAttachmentRendererID() const = 0;
   virtual const FramebufferProps& GetProps() const = 0;
   static Ref<Framebuffer> Create(const FramebufferProps& props);
